# chore(schematics): optimize `getNamedImportReferences` utility